+++ Source/RapyutaSimulationPlugins/Public/Robots/RRBaseRobot.h
@@ -53,6 +53,11 @@ public:
     UPROPERTY(EditAnywhere, BlueprintReadWrite, meta = (DisplayName = "ROS2 Interface Class"))
     TSubclassOf<URRRobotROS2Interface> ROS2InterfaceClass;
 
+    //! Robot's ROS2 Interface
+    UPROPERTY()
+    URRRobotROS2Interface* ROS2Interface = nullptr;
+    void CreateROS2Interface();
+
     /**
      * @brief
      * Actually Object's Name is also unique as noted by UE, but we just do not want to rely on it.
@@ -144,4 +149,10 @@ public:
      */
     // UFUNCTION(BlueprintCallable)
     virtual void SetJointState(const TMap<FString, TArray<float>>& InJointState, const ERRJointControlType InJointControlType);
+
+protected:
+    /**
+     * @brief Instantiate default child components
+     */
+    virtual void PostInitializeComponents() override;
 };